Reorganized marked_tpcf to have clearer control flow
